refactor: rename sheets/ to applications/ and helpers/ to utils/
docs/architecture.md の目指す層構成に合わせてディレクトリ名を揃えた。
import文の書き換えのみで、挙動の変更はない。

以降のリファクタリングでファイルを移動・分割する際に、
パス書き換えの手戻りが混ざらないよう先に済ませておく。
