+++ docs/contributing.md
@@ -77,7 +77,7 @@ feat: add resonance roll dialog
 ## コーディング方針
 
 - TypeScript（strict化を目指す）。型定義の都合で `as any` が残っているが、**新規コードでは増やさない・触った箇所では減らす**
-- UI文字列は `lang/ja.json` が正で、`en.json` はそれに追従する。スキーマの `label` などは `module/helpers/localization.ts` の事前ローカライズ機構を通す
+- UI文字列は `lang/ja.json` が正で、`en.json` はそれに追従する。スキーマの `label` などは `module/utils/localization.ts` の事前ローカライズ機構を通す
 - フォーマット・lintは [Biome](https://biomejs.dev/)（設定: `biome.json`）。手動実行は `npm run check`（修正適用）/ `npm run lint`（検証のみ）
 - 設計の方向性・既知の構造的課題は [アーキテクチャ](/architecture) を参照
