+++ module/applications/types.ts
